Move remaining types
